requirements.txt
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
toml>=0.10.2
jmespath>=1.0.1
jsonschema>=4.19.0
